Split prompts/open-canvas.js into multiple files
Split `prompts/open-canvas.js` into multiple files to improve modularity and maintainability.

* **artifact-meta.js**:
  - Move artifact meta-related constants and schemas.
  - Export `PROGRAMMING_LANGUAGES`, `ARTIFACT_TOOL_SCHEMA`, and `OPTIONALLY_UPDATE_ARTIFACT_META_SCHEMA`.

* **thread-title.js**:
  - Fix newline issue in `TITLE_USER_PROMPT`.

* **open-canvas.js**:
  - Remove artifact meta-related constants and schemas.
  - Import constants and schemas from `artifact-meta.js`, `code-artifact.js`, `followup.js`, `new-artifact.js`, `task.js`, `thread-title.js`, and `user-question.js`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/FLYBYME/canvas-llm?shareId=XXXX-XXXX-XXXX-XXXX).
